Update Neovim and incompatible configs
Updated neovim, and faced some incamplatibilies with which-key
deprecated .register(changed to add) and group and hidden menues.
treesitter was complaining about "install.lua install_info nil" turns
out it was about gotmpl in treesitter config, I had to uncomment the
parser.c stuff to solve it. I had some key binding in barbar related to
which-key, updated them to be compatible with the new version.